import/pyproject.toml
[project]
name = "hagiographies-import"
version = "0.1.0"
description = "Hagiographies data import"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlmodel>=0.0.27",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "sqlalchemy-data-model-visualizer>=0.1.3",
]

[dependency-groups]
dev = ["eralchemy[graphviz]>=1.6.0"]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[project.scripts]
import = "hagiographies_import.cli:main"
